Drone sampling visualization (#55)
* Implement raw keeper primitives

* Implement checking visibility constraints

* Implement publishing jerk control input

* Implement calculating best index for keeper

* Implement calculating best index for keeper

* FIX bug

---------

+++ los_keeper/test/obstacle_manager_test.cc
@@ -46,7 +46,7 @@ TEST_F(ApiTestFixtureObstacle, CheckCollisionBetweenTrajectoryAndObstacles) {
   obstacle_param.planning_horizon = 1.0f;
   obstace_manager = new ObstacleManager(obstacle_param);
   StatePoly trajectory;
-  float time_interval[2]{0.0, obstacle_param.planning_horizon};
+  double time_interval[2]{0.0, obstacle_param.planning_horizon};
   trajectory.SetTimeInterval(time_interval);
   trajectory.SetDegree(3);
   BernsteinCoefficients bernstein_coeff_x{0.0, 1.0, 2.0, 3.0};
